fix react node issue in storage
