+++ src/app/storage.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import { Product } from "@/types/product";
 
 export const collections = [
   {
@@ -46,35 +46,14 @@ export const collections = [
   },
 ];
 
-export const products = [
+export const products: Product[] = [
   {
     id: 1,
     name: "Huy Hiệu Thạch Âm - Xanh Rêu",
     images: ["/products/green-badge.png", "/products/green-badge.jpg"],
     background: "linear-gradient(to right, #e5e5e5, #e9e9e9, #eee)",
     zoom: [],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Huy hiệu này là một biểu tượng tinh tế của dự án Thạch Âm trong môn Đồ
-          án Vốn cổ Khmer, kết hợp giữa yếu tố văn hóa truyền thống và thiết kế
-          hiện đại.
-        </p>
-        <p>
-          Logo "THẠCH ÂM" được trình bày nổi bật, với chữ "THẠCH" nằm gọn trong
-          một khung hình chữ nhật và chữ "ÂM" đặt bên dưới, có thể được tạo hình
-          cách điệu với những đường nét uốn lượn mang đậm phong cách nghệ thuật
-          Khmer cổ điển. Một chi tiết đồ họa uốn lượn lấy cảm hứng từ các hoa
-          văn truyền thống Khmer, có thể là hình ảnh Naga hoặc Kinnari cách
-          điệu, được tích hợp khéo léo vào phần chữ hoặc bao quanh nó, tạo điểm
-          nhấn ấn tượng.
-        </p>
-        <p>
-          Phiên bản xanh rêu: Logo màu vàng đồng hoặc vàng gold nổi bật trên nền
-          xanh rêu, tạo cảm giác sang trọng, cổ kính và quyền quý.
-        </p>
-      </div>
-    ),
+    descriptionKey: "huy-hieu-thach-am-xanh-reu",
     code: "A001_693564536",
     brand: "Thạch Âm",
     material: "Kim loại mạ bóng, in hình phủ nhựa bóng/mờ",
@@ -110,30 +89,8 @@ export const products = [
     images: ["/products/yellow-badge.jpeg", "/products/yellow-badge.png"],
     background: "linear-gradient(to right, #e3e3e5, #e7e7e7, #eaeaea)",
     zoom: [],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Huy hiệu này là một biểu tượng tinh tế của dự án Thạch Âm trong môn Đồ
-          án Vốn cổ Khmer, kết hợp giữa yếu tố văn hóa truyền thống và thiết kế
-          hiện đại.
-        </p>
-        <p>
-          Logo "THẠCH ÂM" được trình bày nổi bật, với chữ "THẠCH" nằm gọn trong
-          một khung hình chữ nhật và chữ "ÂM" đặt bên dưới, có thể được tạo hình
-          cách điệu với những đường nét uốn lượn mang đậm phong cách nghệ thuật
-          Khmer cổ điển. Một chi tiết đồ họa uốn lượn lấy cảm hứng từ các hoa
-          văn truyền thống Khmer, có thể là hình ảnh Naga hoặc Kinnari cách
-          điệu, được tích hợp khéo léo vào phần chữ hoặc bao quanh nó, tạo điểm
-          nhấn ấn tượng.
-        </p>
-        <p>
-          Phiên bản vàng nhạt: Logo màu xanh rêu hoặc đen được đặt trên nền vàng
-          nhạt, mang lại cảm giác nhẹ nhàng, trang nhã nhưng vẫn giữ được nét cổ
-          điển.
-        </p>
-      </div>
-    ),
-    code: "A001_693564536",
+    descriptionKey: "huy-hieu-thach-am-vang-nhat",
+    code: "A001_693564596",
     brand: "Thạch Âm",
     material: "Kim loại mạ bóng, in hình phủ nhựa bóng/mờ",
     pin: "Ghim kim cài lưng (an toàn, dễ sử dụng)",
@@ -168,52 +125,8 @@ export const products = [
     images: ["/products/bandana.png"],
     background: "linear-gradient(to right, #ebebeb, #eee, #ebebeb)",
     zoom: [],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>"Khăn Bandana Thạch Âm: Nét Vàng Cổ Tích Giữa Làn Gió Hiện Đại."</p>
-        <p>
-          Hãy tưởng tượng bạn đang lạc bước giữa những hành lang đá rêu phong
-          của Angkor, ánh nắng chiều hắt lên những phù điêu vàng son. Đó chính
-          là cảm hứng bất tận mà nhóm Thạch Âm đã chắt lọc, gửi gắm vào từng sợi
-          vải của chiếc khăn bandana đặc biệt này.
-        </p>
-        <p>
-          Với sắc xanh rêu trầm mặc như thời gian ngủ yên trên đá, và những họa
-          tiết vàng đồng tinh xảo như chính những kho báu ẩn mình, Khăn Bandana
-          "Dấu Vàng Thời Gian" là sự giao thoa hoàn hảo giữa quá khứ lẫy lừng và
-          phong cách cá nhân độc đáo. Mỗi chiếc khăn là một bản tuyên ngôn về vẻ
-          đẹp di sản, được thiết kế thủ công đầy tâm huyết trong khuôn khổ môn
-          Đồ án Vốn cổ Khmer.
-        </p>
-        <p>
-          Hãy để chiếc khăn này không chỉ là phụ kiện, mà là câu chuyện bạn mang
-          theo, là dấu ấn của một nền văn minh rực rỡ, sẵn sàng cùng bạn khám
-          phá những hành trình mới.
-        </p>
-      </div>
-    ),
-    instruction: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Để giữ gìn vẻ đẹp nguyên bản và giá trị tinh thần của Khăn Bandana
-          Thạch Âm, bạn nên giặt tay nhẹ nhàng với nước lạnh và xà phòng dịu
-          nhẹ.
-        </p>
-        <p>
-          Tránh sử dụng thuốc tẩy hoặc giặt máy ở chế độ mạnh, vì có thể làm
-          phai màu và ảnh hưởng đến chất vải thủ công.
-        </p>
-        <p>
-          Khi phơi, nên phơi khăn ở nơi thoáng mát, tránh ánh nắng trực tiếp để
-          giữ cho sắc xanh rêu và họa tiết vàng đồng luôn rực rỡ theo thời gian.
-        </p>
-        <p>
-          Khi không sử dụng, bạn có thể gấp gọn khăn và bảo quản trong túi vải
-          hoặc hộp, như cách lưu giữ một mảnh di sản - để mỗi lần mở ra, vẫn vẹn
-          nguyên cảm xúc ban đầu.
-        </p>
-      </div>
-    ),
+    descriptionKey: "khan-bandana-thach-am",
+    instructionKey: "khan-bandana-thach-am",
     code: "A001_694535520",
     brand: "Thạch Âm",
     material: "Lụa cao cấp",
@@ -244,68 +157,7 @@ export const products = [
     ],
     background: "#797979",
     zoom: [2, 3],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>Giới Thiệu Bộ Card Visit "Thạch Âm: Dấu Ấn Di Sản Khmer"</p>
-        <p>"Không chỉ là card visit, đây là một tuyên ngôn văn hóa."</p>
-        <p>
-          Trong thế giới hiện đại, nơi sự kết nối là chìa khóa, nhóm Thạch Âm tự
-          hào giới thiệu bộ card visit độc đáo, tinh hoa của quá trình nghiên
-          cứu và sáng tạo trong môn Đồ án Vốn cổ Khmer. Mỗi tấm card không chỉ
-          mang thông tin liên hệ, mà còn là một tác phẩm nghệ thuật, một minh
-          chứng sống động cho vẻ đẹp trường tồn của di sản Khmer.
-        </p>
-        <div className="space-y-2">
-          <p>Thiết kế song hành - Vẻ đẹp đối lập hài hòa:</p>
-          <p>
-            Mặt A (Xanh Rêu Đậm & Vàng Đồng): Nổi bật với tông màu xanh rêu sâu
-            thẳm, gợi nhớ đến những phiến đá cổ kính rêu phong và khu rừng huyền
-            bí bao quanh các đền đài Angkor. Trên nền xanh ấy, logo "THẠCH ÂM"
-            và các họa tiết Khmer cổ điển được in nổi hoặc dập chìm với sắc vàng
-            đồng rực rỡ, tượng trưng cho sự thịnh vượng, quyền quý và ánh hào
-            quang của một nền văn minh vàng son. Mặt này thể hiện sự vững chãi,
-            uy nghi và chiều sâu văn hóa.
-          </p>
-          <p>
-            Mặt B (Vàng Nhạt & Xanh Rêu Đậm): Mang một sắc thái tươi sáng hơn
-            với nền vàng nhạt (kem/ngà), tượng trưng cho sự tinh khiết, ánh sáng
-            và sự khởi đầu mới. Các họa tiết Khmer uốn lượn tinh tế và thông tin
-            liên hệ được trình bày bằng màu xanh rêu đậm, tạo sự tương phản đầy
-            cuốn hút. Mặt này thể hiện sự thanh lịch, gần gũi và tính ứng dụng
-            hiện đại, nhưng vẫn giữ được nét cổ kính qua các chi tiết hoa văn.
-          </p>
-        </div>
-        <div className="space-y-2">
-          <p>Điểm nhấn từ họa tiết:</p>
-          <p>
-            Mỗi họa tiết trên card được nhóm Thạch Âm tỉ mỉ nghiên cứu và cách
-            điệu từ những hoa văn truyền thống Khmer, như hình ảnh Naga uốn
-            lượn, các đường kỷ hà phức tạp, hay những mô-típ lá cây, hoa sen
-            quen thuộc trong kiến trúc và điêu khắc cổ. Điều này không chỉ tăng
-            tính thẩm mỹ mà còn kể một câu chuyện về bản sắc văn hóa.
-          </p>
-        </div>
-        <div className="space-y-2">
-          <p>Chất liệu và hoàn thiện:</p>
-          <p>
-            Bộ card được in trên chất liệu giấy cao cấp, có độ dày dặn và bề mặt
-            được xử lý tinh tế (giấy mỹ thuật có cán mờ/bóng nhẹ), mang lại cảm
-            giác sang trọng và đẳng cấp khi cầm trên tay.
-          </p>
-        </div>
-        <div className="space-y-2">
-          <p>Ý nghĩa:</p>
-          <p>
-            Bộ card visit "Thạch Âm: Dấu Ấn Di Sản Khmer" không chỉ là công cụ
-            giao tiếp. Đó là một lời mời khám phá, một cầu nối giữa quá khứ vĩ
-            đại và hiện tại đầy tiềm năng, là biểu tượng cho sự sáng tạo và tâm
-            huyết của nhóm Thạch Âm trong việc gìn giữ và lan tỏa giá trị vốn cổ
-            Khmer.
-          </p>
-        </div>
-        <p>Hãy trao đi tấm card này, và bạn đang trao đi một câu chuyện.</p>
-      </div>
-    ),
+    descriptionKey: "card-visit-thach-am",
     code: "A001_01134034503",
     brand: "Thạch Âm",
     material: "Giấy mỹ thuật cao cấp",
@@ -331,102 +183,9 @@ export const products = [
     images: ["/products/hand-fan-1.jpeg"],
     background: "#FCFCFC",
     zoom: [],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>Quạt Tay Thạch Âm - Nét Đẹp Huyền Bí Từ Họa Tiết Khmer</p>
-        <p>
-          Chào mừng bạn đến với Quạt Tay Thạch Âm – một tác phẩm nghệ thuật thủ
-          công tinh xảo, nơi vẻ đẹp truyền thống giao thoa cùng sự thanh lịch
-          hiện đại.
-        </p>
-        <p>
-          Lấy cảm hứng từ những họa tiết Khmer cổ kính, mỗi chiếc quạt Thạch Âm
-          không chỉ là một vật dụng làm mát mà còn là một câu chuyện về văn hóa
-          và lịch sử. Từng đường nét, họa tiết được thể hiện một cách tỉ mỉ, gợi
-          lên sự uy nghi của các vị thần, sự thanh tịnh của Đức Phật và sự huyền
-          bí của những ngôi đền cổ.
-        </p>
-        <p>
-          Sản phẩm được chế tác từ chất liệu giấy cao cấp, mang lại cảm giác nhẹ
-          nhàng, êm ái khi sử dụng. Đặc biệt, với độ bền cao và khả năng giữ màu
-          tốt, Quạt Tay Thạch Âm sẽ là một điểm nhấn độc đáo cho không gian sống
-          của bạn, hoặc là món quà ý nghĩa dành tặng cho những người thân yêu.
-        </p>
-        <div className="space-y-2">
-          <p>Điểm nổi bật của Quạt Tay Thạch Âm:</p>
-          <p>
-            Thiết kế độc đáo: Ứng dụng tinh hoa họa tiết Khmer, mang đậm bản sắc
-            văn hóa.
-          </p>
-          <p>Chất liệu cao cấp: Giấy bền đẹp, thân thiện với môi trường.</p>
-          <p>
-            Giá trị nghệ thuật: Không chỉ là quạt, mà còn là một tác phẩm trang
-            trí tinh tế.
-          </p>
-          <p>
-            Món quà ý nghĩa: Biểu tượng của sự may mắn, bình an và vẻ đẹp truyền
-            thống
-          </p>
-        </div>
-        <div className="space-y-2">
-          <p>Ứng dụng:</p>
-          <p>✔️ Quà tặng sự kiện văn hóa – lễ hội truyền thống</p>
-          <p>✔️ Vật phẩm lưu niệm cho khách du lịch</p>
-          <p>✔️ Phụ kiện trình diễn nghệ thuật hoặc trưng bày trang trí</p>
-          <p>✔️ Quà tặng doanh nghiệp mang bản sắc văn hóa</p>
-        </div>
-        <p>
-          Hãy để Quạt Tay Thạch Âm mang đến cho bạn không chỉ làn gió mát lành
-          mà còn là cảm giác bình yên, thư thái và sự kết nối với những giá trị
-          văn hóa ngàn đời.
-        </p>
-      </div>
-    ),
-    instruction: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Tránh nước và độ ẩm: Quạt được làm từ chất liệu giấy, do đó tuyệt đối
-          tránh tiếp xúc trực tiếp với nước, mưa hoặc môi trường có độ ẩm cao.
-          Nước có thể làm nhăn, phai màu hoặc làm hỏng cấu trúc giấy.
-        </p>
-        <p>
-          Tránh ánh nắng trực tiếp và nhiệt độ cao: Không để quạt dưới ánh nắng
-          mặt trời trực tiếp trong thời gian dài hoặc gần các nguồn nhiệt (như
-          lò sưởi, bếp nóng). Nhiệt độ cao có thể làm giấy khô, giòn và phai
-          màu.
-        </p>
-        <p>
-          Bảo quản nơi khô ráo, thoáng mát: Khi không sử dụng, nên cất giữ quạt
-          ở nơi khô ráo, thoáng mát, tránh bụi bẩn và các vật nặng đè lên.
-        </p>
-        <p>
-          Tránh va đập và vật sắc nhọn: Không làm rơi quạt hoặc để quạt va chạm
-          với các vật cứng, sắc nhọn có thể làm rách giấy hoặc gãy nan.
-        </p>
-        <p>
-          Vệ sinh: Để vệ sinh quạt, bạn có thể dùng cọ mềm hoặc khăn khô, sạch
-          để nhẹ nhàng lau bỏ bụi bẩn trên bề mặt. Không dùng khăn ướt hoặc hóa
-          chất tẩy rửa
-        </p>
-      </div>
-    ),
-    note: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Quạt được thiết kế để sử dụng nhẹ nhàng. Tránh vung quạt quá mạnh hoặc
-          dùng lực tác động lên các nan quạt khi đang mở.
-        </p>
-        <p>
-          Sản phẩm không dành cho trẻ em dưới 3 tuổi sử dụng mà không có sự giám
-          sát của người lớn.
-        </p>
-        <p>
-          Do đặc tính của chất liệu giấy và kỹ thuật thủ công, mỗi chiếc quạt có
-          thể có sự khác biệt nhỏ về màu sắc hoặc chi tiết hoa văn, tạo nên nét
-          độc đáo riêng cho từng sản phẩm.
-        </p>
-      </div>
-    ),
+    descriptionKey: "quat-tay-thach-am-khmer-wind",
+    instructionKey: "quat-tay-thach-am-khmer-wind",
+    noteKey: "quat-tay-thach-am-khmer-wind",
     code: "A001_01134034745",
     brand: "Thạch Âm",
     material: "Giấy cao cấp",
@@ -465,103 +224,9 @@ export const products = [
     images: ["/products/hand-fan-2.jpg"],
     background: "#FCFCFC",
     zoom: [],
-    description: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>Quạt Tay Thạch Âm - Nét Đẹp Huyền Bí Từ Họa Tiết Khmer</p>
-        <p>
-          Chào mừng bạn đến với Quạt Tay Thạch Âm – một tác phẩm nghệ thuật thủ
-          công tinh xảo, nơi vẻ đẹp truyền thống giao thoa cùng sự thanh lịch
-          hiện đại.
-        </p>
-        <p>
-          Chiếc quạt giấy "Naga Wind" là một tác phẩm nghệ thuật thu nhỏ, kết
-          tinh từ nét đẹp truyền thống Khmer và tinh thần đương đại của thương
-          hiệu Thạch Âm. Lấy cảm hứng từ họa tiết linh vật, chim thần, hoa sen,
-          và chùa Khmer cổ, chiếc quạt không chỉ là vật phẩm thủ công mà còn là
-          biểu tượng của di sản văn hóa và âm thanh linh thiêng.
-        </p>
-        <p>
-          “Naga Wind” không đơn thuần là một chiếc quạt, mà là làn gió mang theo
-          thanh âm di sản – nơi nghệ thuật và tâm linh Khmer hội tụ trong từng
-          đường nét. Đây là lời tri ân sâu sắc đến những giá trị cổ truyền của
-          người Khmer Nam Bộ, đặc biệt là vùng Trà Vinh – nơi khởi nguồn thương
-          hiệu Thạch Âm.
-        </p>
-        <div className="space-y-2">
-          <p>Điểm nổi bật của Quạt Tay Thạch Âm:</p>
-          <p>
-            Thiết kế độc đáo: Ứng dụng tinh hoa họa tiết Khmer, mang đậm bản sắc
-            văn hóa.
-          </p>
-          <p>Chất liệu cao cấp: Giấy bền đẹp, thân thiện với môi trường.</p>
-          <p>
-            Giá trị nghệ thuật: Không chỉ là quạt, mà còn là một tác phẩm trang
-            trí tinh tế.
-          </p>
-          <p>
-            Món quà ý nghĩa: Biểu tượng của sự may mắn, bình an và vẻ đẹp truyền
-            thống
-          </p>
-        </div>
-        <div className="space-y-2">
-          <p>Ứng dụng:</p>
-          <p>✔️ Quà tặng sự kiện văn hóa – lễ hội truyền thống</p>
-          <p>✔️ Vật phẩm lưu niệm cho khách du lịch</p>
-          <p>✔️ Phụ kiện trình diễn nghệ thuật hoặc trưng bày trang trí</p>
-          <p>✔️ Quà tặng doanh nghiệp mang bản sắc văn hóa</p>
-        </div>
-        <p>
-          Hãy để Quạt Tay Thạch Âm mang đến cho bạn không chỉ làn gió mát lành
-          mà còn là cảm giác bình yên, thư thái và sự kết nối với những giá trị
-          văn hóa ngàn đời.
-        </p>
-      </div>
-    ),
-    instruction: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Tránh nước và độ ẩm: Quạt được làm từ chất liệu giấy, do đó tuyệt đối
-          tránh tiếp xúc trực tiếp với nước, mưa hoặc môi trường có độ ẩm cao.
-          Nước có thể làm nhăn, phai màu hoặc làm hỏng cấu trúc giấy.
-        </p>
-        <p>
-          Tránh ánh nắng trực tiếp và nhiệt độ cao: Không để quạt dưới ánh nắng
-          mặt trời trực tiếp trong thời gian dài hoặc gần các nguồn nhiệt (như
-          lò sưởi, bếp nóng). Nhiệt độ cao có thể làm giấy khô, giòn và phai
-          màu.
-        </p>
-        <p>
-          Bảo quản nơi khô ráo, thoáng mát: Khi không sử dụng, nên cất giữ quạt
-          ở nơi khô ráo, thoáng mát, tránh bụi bẩn và các vật nặng đè lên.
-        </p>
-        <p>
-          Tránh va đập và vật sắc nhọn: Không làm rơi quạt hoặc để quạt va chạm
-          với các vật cứng, sắc nhọn có thể làm rách giấy hoặc gãy nan.
-        </p>
-        <p>
-          Vệ sinh: Để vệ sinh quạt, bạn có thể dùng cọ mềm hoặc khăn khô, sạch
-          để nhẹ nhàng lau bỏ bụi bẩn trên bề mặt. Không dùng khăn ướt hoặc hóa
-          chất tẩy rửa
-        </p>
-      </div>
-    ),
-    note: (
-      <div className="space-y-4 text-sm tracking-normal text-gray-700 md:tracking-wide">
-        <p>
-          Quạt được thiết kế để sử dụng nhẹ nhàng. Tránh vung quạt quá mạnh hoặc
-          dùng lực tác động lên các nan quạt khi đang mở.
-        </p>
-        <p>
-          Sản phẩm không dành cho trẻ em dưới 3 tuổi sử dụng mà không có sự giám
-          sát của người lớn.
-        </p>
-        <p>
-          Do đặc tính của chất liệu giấy và kỹ thuật thủ công, mỗi chiếc quạt có
-          thể có sự khác biệt nhỏ về màu sắc hoặc chi tiết hoa văn, tạo nên nét
-          độc đáo riêng cho từng sản phẩm.
-        </p>
-      </div>
-    ),
+    descriptionKey: "quat-tay-thach-am-naga-wind",
+    instructionKey: "quat-tay-thach-am-naga-wind",
+    noteKey: "quat-tay-thach-am-naga-wind",
     code: "A001_01134034766",
     brand: "Thạch Âm",
     material: "Giấy cao cấp",
